requirements.txt
# requirements.txtstreamlityt-dlpffmpeg-pythongoogle-generativeaifpdf2